added support for 2 new flags `ignoreType` & `interval`
`--ignoreType` allows a type to be ignored from the dump/load. Three options are supported. `data,mapping,analyzer`.
Multi-type support is available, when used each type must be comma(,)-separated
`interval` allows control over the interval for spawning a dump/load for a new index.
For small indices this can be set to `0` to reduce delays and optimize performance